+++ examples/websocket-chat-ts-run/game-server/app.ts
@@ -1,4 +1,12 @@
-import { FrontendOrBackendSession, HandlerCallback, pinus, RESERVED, RouteRecord } from 'pinus';
+import {
+    createTcpAcceptor,
+    createTcpMailBox,
+    FrontendOrBackendSession,
+    HandlerCallback,
+    pinus,
+    RESERVED,
+    RouteRecord
+} from 'pinus';
 import './app/servers/user.rpc.define'
 import * as  routeUtil from './app/util/routeUtil';
 import { preload } from './preload';
@@ -43,6 +51,11 @@ app.configure('production|development', 'connector', function () {
             useDict: true,
             useProtobuf: true
         });
+    // 不自动按照路由生成router,仅使用 config/dictionary 内的路由.
+    // 具体看 packages/pinus/lib/components/dictionary.ts DictionaryComponentOptions
+    app.set('dictionaryConfig', {
+        ignoreAutoRouter: true,
+    })
 
     /**
      // 缓存大小不够 日志示例
@@ -113,6 +126,20 @@ app.configure('production|development', function () {
 
     // filter configures
     app.filter(new pinus.filters.timeout());
+
+    // RPC 启用TCP协议
+    app.set('proxyConfig', {
+        mailboxFactory: createTcpMailBox,
+        //    bufferMsg:true
+        // rpc 超时时间
+        // timeout: 20 * 1000,
+        // dynamicUserProxy: true,
+    });
+    app.set('remoteConfig', {
+        acceptorFactory: createTcpAcceptor,
+        // bufferMsg:true,
+        // interval:50,
+    });
 });
 
 app.configure('development', function () {