+++ packages/pinus/test/service/connectionService.ts
@@ -134,8 +134,8 @@ describe('connection service test', function () {
         let infos = sinfo.loginedList;
         should.exist(infos);
         infos.length.should.equal(2);
-        infos.should.include(info1);
-        infos.should.include(info2);
+        infos.should.containEql(info1);
+        infos.should.containEql(info2);
 
         done();
     });