+++ packages/pinus/test/util/pathUtil.ts
@@ -62,7 +62,7 @@ describe('path util test', function () {
             should.exist(p);
             expectNames.length.should.equal(p.length);
             for (let i = 0, l = expectNames.length; i < l; i++) {
-                p.should.include(expectNames[i]);
+                p.should.containEql(expectNames[i]);
             }
         });